+++ tests/blocks_data_source/main.tf
@@ -1,4 +1,4 @@
-# Returns all spaces with network and usage information
+# Returns all blocks with network and usage information
 data "azureipam_blocks" "expanded" {
   space              = "au"
   expand             = true
@@ -9,12 +9,12 @@ output "expanded_blocks" {
 }
 
 
-# Returns all spaces without network and usage information
+# Returns all blocks without network and usage information
 data "azureipam_blocks" "not_expanded" {
   space              = "au"
   expand             = false
   append_utilization = false
 }
-output "not_expanded_spaces" {
+output "not_expanded_blocks" {
   value = data.azureipam_blocks.not_expanded
 }