+++ CHANGELOG.md
@@ -44,6 +44,8 @@
 + data resource `azureipam_blocks` to get a list of all external networks associated with a space and block. 
 + resource `azureipam_reservation_cidr` to allow to create reservation specifying the cidr, space and block.
 + data resource `azureipam_external` to get a specific external network associated with a space and block. 
++ data resource `azureipam_block` to get a specific block with related information.
+
 
 ### Modified (Breaking Change)
 + resource `azureipam_reservation` now allow to specify a block list. The list is evaluated in the order provider